Define a basic app the can serialize a Kickstart file
This required defining the Kickstart model and KickstartSerializer.
This was pretty straight forward, but the collection check needs to
be disabled as `Struct` and therefore `Kickstart` responds to each.
This does not make it a collection however, as actually means each key
not each model

There was a bit of weirdness with requiring the Accept and Content-Type
headers to be set correctly. This is because sinja preforms a strict
content type check. Our does not require this check, so it has been
"disabled" using some middleware. In actuality it just sets the
headers correctly for all requests