Start working on the upload routes
There was a bit of an issue with `sinja` only accepting JSON api inputs.
This breaks the file upload as it requires `application/octet-stream`

A work around is to use some middleware to cache the original body
elsewhere on the `env`. This allows the actual request body to be
reset to empty string. This prevents sinja from trying to parse the
stream.

The rest of the `Upload` stuff is experimental and might change