Resolve models from their filename
Models with a single id input can be resolved easily from their
filename. This requires constructing a regex match that captures all the
arguments used to generate the filename. This can then be used to
reverse identify the original model.

A SingleIDFileModel abstract class has been added, as a lot but not all
the models take a single id input