Determine the current index based on directories
Previously it was based on whether the include subnet file exist.
However this may lead to weird bugs if the file doesn't exist (such as
when the app is first installed).

Instead the existence is based on the directory present or not. The
specs have been updated to reflect this change