Safely use `Dir.rmdir` when possible
Some of the time, the directory being removed should be empty. If it
isn't then some edge case has been triggered and the app should fail.

However this is always paired with a full directory of files, which
should be removed.

On success:
tmp: The original files can be safely removed with `rm_rf`
old: Should be an empty directory `rmdir`

On failure:
tmp: Old files should have been restored, remove with `rmdir`
new: Can be safely removed with `rm_rf`