Do not store the DHCP configs with auto index
The auto incrementing index means the files only work with relative
paths. However relative paths are worked out from the current working
directory, which is unpredictable and out of MetalServer control

Instead a static `current` directory is always used to store the
configs, this allows absolute paths to be used.

The old files are still backed up to a temporary directory to allow for
fallback.

As the paths are now static, their is no longer a need for a master
DHCP config