Patch FlightConfig::Globber to escape paths
Previously `.` in the file extension acted as wild cards. This creates a
bunch of edge cases where the match gets confused. The solution is to
escape the static file path components, and only match the argument
section.

This requires a sub to set the match regex, otherwise it would also be
escaped